Some useful loon histogram functions

Here list several useful functions for loon histograms.

getBinData(widget)

l_getBinIds(widget)

l_breaks(widget)

l_binCut(widget, labels = NULL, digits = 2)

Arguments

widget
A loon l_hist widget.

labels
Labels for the levels of the resulting category. By default, labels are constructed using "(a,b]" interval notation. If labels = FALSE, simple integer codes are returned instead of a factor. If labels are recognized as colors, character will be returned directly.

digits
The number of digits used in formatting the breaks.

Details

getBinData is utilized to extract loon bin information, such as the counts, locations, points id of each colored bin. All rest useful functions in this page are built based on it.

l_getBinIds returns the ids of each bin

l_breaks returns the range (breaks) of each bin

l_binCut divides l_hist widget x into current histogram intervals and codes values x according to which interval they fall. It is very similar to cut in base package.

Examples

# getBinData
p <- l_hist(mtcars$mpg)
getBinData(p)

#> $`bin 1`
#> $`bin 1`$count
#> $`bin 1`$count$all
#> [1] 11
#> 
#> $`bin 1`$count$selected
#> [1] 0
#> 
#> $`bin 1`$count$`#999999999999`
#> [1] 11
#> 
#> 
#> $`bin 1`$points
#> $`bin 1`$points$all
#>  [1]  7 12 14 15 16 17 22 23 24 29 31
#> 
#> $`bin 1`$points$selected
#> numeric(0)
#> 
#> $`bin 1`$points$`#999999999999`
#>  [1]  7 12 14 15 16 17 22 23 24 29 31
#> 
#> 
#> $`bin 1`$x0
#> [1] 10.4
#> 
#> $`bin 1`$x1
#> [1] 17.02531
#> 
#> 
#> $`bin 2`
#> $`bin 2`$count
#> $`bin 2`$count$all
#> [1] 14
#> 
#> $`bin 2`$count$selected
#> [1] 0
#> 
#> $`bin 2`$count$`#999999999999`
#> [1] 14
#> 
#> 
#> $`bin 2`$points
#> $`bin 2`$points$all
#>  [1]  1  2  3  4  5  6  9 10 11 13 21 25 30 32
#> 
#> $`bin 2`$points$selected
#> numeric(0)
#> 
#> $`bin 2`$points$`#999999999999`
#>  [1]  1  2  3  4  5  6  9 10 11 13 21 25 30 32
#> 
#> 
#> $`bin 2`$x0
#> [1] 17.02531
#> 
#> $`bin 2`$x1
#> [1] 23.65062
#> 
#> 
#> $`bin 3`
#> $`bin 3`$count
#> $`bin 3`$count$all
#> [1] 3
#> 
#> $`bin 3`$count$selected
#> [1] 0
#> 
#> $`bin 3`$count$`#999999999999`
#> [1] 3
#> 
#> 
#> $`bin 3`$points
#> $`bin 3`$points$all
#> [1]  8 26 27
#> 
#> $`bin 3`$points$selected
#> numeric(0)
#> 
#> $`bin 3`$points$`#999999999999`
#> [1]  8 26 27
#> 
#> 
#> $`bin 3`$x0
#> [1] 23.65062
#> 
#> $`bin 3`$x1
#> [1] 30.27593
#> 
#> 
#> $`bin 4`
#> $`bin 4`$count
#> $`bin 4`$count$all
#> [1] 4
#> 
#> $`bin 4`$count$selected
#> [1] 0
#> 
#> $`bin 4`$count$`#999999999999`
#> [1] 4
#> 
#> 
#> $`bin 4`$points
#> $`bin 4`$points$all
#> [1] 18 19 20 28
#> 
#> $`bin 4`$points$selected
#> numeric(0)
#> 
#> $`bin 4`$points$`#999999999999`
#> [1] 18 19 20 28
#> 
#> 
#> $`bin 4`$x0
#> [1] 30.27593
#> 
#> $`bin 4`$x1
#> [1] 36.90124
#> 
#> 

# l_binCut
p <- l_hist(iris)
binCut <- l_binCut(p)
p['color'] <- binCut
## ggplot color hue
gg_color_hue <- function(n) {
  hues <- seq(15, 375, length = n + 1)
  hcl(h = hues, l = 65, c = 100)[1:n]
}

# l_getBinIds
## number of bins
binIds <- l_getBinIds(p)
p['color'] <- l_binCut(p, labels = gg_color_hue(length(binIds)))

## The count of each bin
lengths(binIds)

#> bin 1 bin 2 bin 3 bin 4 bin 5 bin 6 bin 7 
#>    16    30    37    32    23     6     6 

# l_breaks
breaks <- l_breaks(p)