Create and optionally draw a grid grob from a loon widget handle

Create and optionally draw a grid grob from a loon widget handle

grid.loon(target, name = NULL, gp = gpar(), draw = TRUE, vp = NULL)

Arguments

target
either an object of class loon or a vector that specifies the widget, layer, glyph, navigator or context completely. The widget is specified by the widget path name (e.g. '.l0.plot'), the remaining objects by their ids.

name
a character identifier for the grob, or NULL. Used to find the grob on the display list and/or as a child of another grob.

gp
a gpar object, or NULL, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

draw
a logical value indicating whether graphics output should be produced.

vp
a grid viewport object (or NULL).

Value

a grid grob of the loon plot

See also

Examples

if (FALSE) {
library(grid)
widget <- with(iris, l_plot(Sepal.Length, Sepal.Width))
grid.loon(widget)
}