Query the aspect ratio of a plot

The aspect ratio is defined by the ratio of the number of pixels for one data unit on the y axis and the number of pixels for one data unit on the x axes.

l_aspect(widget)

Arguments

widget
widget path as a string or as an object handle

Value

aspect ratio

Examples

p <- with(iris, l_plot(Sepal.Length ~ Sepal.Width, color=Species))

l_aspect(p)

#> [1] 0.4562569

l_aspect(p) <- 1