Get the event pattern and callback Tcl code of a canvas binding

This function returns the registered event pattern and the Tcl callback code that the Tcl interpreter evaluates after a event occurs that machtches the event pattern.

l_bind_canvas_get(widget, id)

Arguments

widget
widget path as a string or as an object handle

id
canvas binding id

Value

Character vector of length two. First element is the event pattern, the second element is the Tcl callback code.

Details

Bindings, callbacks, and binding substitutions are described in detail in loon's documentation webpage, i.e. run l_help("learn_R_bind")

See also

Examples

# binding for when plot is resized 
p <- l_plot(iris[,1:2], color=iris$Species)

printSize <- function(p) {
    size <- l_size(p)
    cat(paste('Size of widget ', p, ' is: ',
              size[1], 'x', size[2], ' pixels\n', sep=''))
}

l_bind_canvas(p, event='<Configure>', function(W) {printSize(W)})

#> [1] "canvasBinding0"

id <- l_bind_canvas_ids(p)
id

#> [1] "canvasBinding0"

l_bind_canvas_get(p, id)

#> [1] "<Configure>"              "R_call 0x7fd6a8e3baa8 %W"