Reorder the state binding evaluation sequence

The order the state bindings defines how they get evaluated once an event matches event patterns of multiple state bindings.

l_bind_state_reorder(target, ids)

Arguments

target
either an object of class loon or a vector that specifies the widget, layer, glyph, navigator or context completely. The widget is specified by the widget path name (e.g. '.l0.plot'), the remaining objects by their ids.

ids
new state binding id evaluation order, this must be a rearrangement of the elements returned by the l_bind_state_ids function.

Value

vector with binding id evaluation order (same as the id argument)

Details

Bindings, callbacks, and binding substitutions are described in detail in loon's documentation webpage, i.e. run l_help("learn_R_bind")

See also