Query a Plot State

All of loon's displays have plot states. Plot states specify what is displayed, how it is displayed and if and how the plot is linked with other loon plots. Layers, glyphs, navigators and contexts have states too (also refered to as plot states). This function queries a single plot state.

l_cget(target, state)

Arguments

target
either an object of class loon or a vector that specifies the widget, layer, glyph, navigator or context completely. The widget is specified by the widget path name (e.g. '.l0.plot'), the remaining objects by their ids.

state
state name

See also

Examples

p <- l_plot(iris, color = iris$Species)

#> Warning: ‘>’ not meaningful for factors

l_cget(p, "color")

#>   [1] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>   [5] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>   [9] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [13] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [17] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [21] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [25] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [29] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [33] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [37] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [41] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [45] "#999999999999" "#999999999999" "#999999999999" "#999999999999"
#>  [49] "#999999999999" "#999999999999" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [53] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [57] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [61] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [65] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [69] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [73] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [77] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [81] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [85] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [89] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [93] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#>  [97] "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3" "#A6A6CECEE3E3"
#> [101] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [105] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [109] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [113] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [117] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [121] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [125] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [129] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [133] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [137] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [141] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [145] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"
#> [149] "#FFFFC0C0CBCB" "#FFFFC0C0CBCB"

p['selected']

#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE