Export a loon plot as an image

The supported image formats are dependent on the system environment. Plots can always be exported to the Postscript format. Exporting displays as .pdfs is only possible when the command line tool epstopdf is installed. Finally, exporting to either png, jpg, bmp, tiff or gif requires the Img Tcl extension. When choosing one of the formats that depend on the Img extension, it is possible to export any Tk widget as an image including inspectors.

l_export(widget, filename, width, height)

Arguments

widget
widget path as a string or as an object handle

filename
path of output file

width
image width in pixels

height
image height in pixels

Value

path to the exported file

Details

Note that the CTRL-T key combination opens a dialog to export he graphic.

The native export format is to ps as this is what the Tk canvas offers. If the the l_export fails with other formats then please resort to a screen capture method for the moment.

See also