Add a Serialaxes Glyph

Serialaxes glyph show either a star glyph or a parallel coordinate glyph for each point.

Usage

l_glyph_add_serialaxes(
  widget,
  data,
  sequence,
  linewidth = 1,
  scaling = "variable",
  axesLayout = "radial",
  showAxes = FALSE,
  andrews = FALSE,
  axesColor = "gray70",
  showEnclosing = FALSE,
  bboxColor = "gray70",
  label = "",
  ...
)

Arguments

widget

widget path as a string or as an object handle

data

a data frame with numerical data only

sequence

vector with variable names that defines the axes sequence

linewidth

linewidth of outline

scaling

one of 'variable', 'data', 'observation' or 'none' to specify how the data is scaled. See Details and Examples for more information.

axesLayout

either "radial" or "parallel"

showAxes

boolean to indicate whether axes should be shown or not

andrews

Andrew's curve (a 'Fourier' transformation)

axesColor

color of axes

showEnclosing

boolean, circle (axesLayout=radial) or sqaure (axesLayout=parallel) to show bounding box/circle of the glyph (or showing unit circle or rectangle with height 1 if scaling=none)

bboxColor

color of bounding box/circle

label

label of a glyph (currently shown only in the glyph inspector)

...

state arguments

See also

Examples

if(interactive()){

p <- with(olive, l_plot(oleic, stearic, color=Area))
gs <- l_glyph_add_serialaxes(p, data=olive[,-c(1,2)], showArea=FALSE)
p['glyph'] <- gs
}