Generic funtction to create an interactive graph display

Interactive graphs in loon are currently most often used for navigation graphs.

Usage

l_graph(nodes, ...)

# S3 method for class 'graph'
l_graph(nodes, ...)

# S3 method for class 'loongraph'
l_graph(nodes, ...)

# Default S3 method
l_graph(nodes = "", from = "", to = "", isDirected = FALSE, parent = NULL, ...)

Arguments

nodes

object for method dispatch

...

arguments passed on to methods

from

vector with node names of the from-to pairs for edges

to

vector with node names of the from-to pairs for edges

isDirected

a boolean state to specify whether these edges have directions

parent

parent widget of graph display

Value

graph handle

Details

For more information run: l_help("learn_R_display_graph.html#graph")

See also

Examples

if(interactive()) {
 G <- completegraph(nodes=names(iris))
 LG <- linegraph(G, sep=":")
 g <- l_graph(LG)
}