Add a graph to the graphswitch widget using a graph object

Graph objects are defined in the graph R package.

Usage

# S3 method for class 'graph'
l_graphswitch_add(widget, graph, label = "", index = "end", ...)

Arguments

widget

graphswitch widget handle (or widget path)

graph

a graph object created with the functions in the graph R package.

label

string with label for graph

index

position of graph in the graph list

...

additional arguments are not used for this methiod

Value

id for graph in the graphswitch widget

See also