Import Greyscale Images as Tcl images from an Array

Import image grayscale data (0-255) with each image saved as a row or column of an array.

Usage

l_image_import_array(
  array,
  width,
  height,
  img_in_row = TRUE,
  invert = FALSE,
  rotate = 0
)

Arguments

array

of 0-255 grayscale value data.

width

of images in pixels.

height

of images in pixels.

img_in_row

logical, TRUE if every row of the array represents an image

invert

logical, for 'invert=FALSE' 0=withe, for 'invert=TRUE' 0=black

rotate

the image: one of 0, 90, 180, or 270 degrees.

Value

vector of image object names

Details

Images in tcl are managed by the tcl interpreter and made accessible to the user via a handle, i.e. a function name of the form image1, image2, etc.

For more information run: l_help("learn_R_display_plot.html#images")

Examples

if (FALSE) { # \dontrun{
# see
demo("l_ng_images_frey_LLE")
} # }