Layer line in Line object

Methods to plot map data defined in the sp package

# S3 method for Line
l_layer(widget, x, ...)

Arguments

widget
widget widget path as a string or as an object handle

x
an object defined in the sp class

...
arguments forwarded to the relative l_layer function

Value

layer id

Details

Note that currently loon does neither support holes and ring directions.

References

Applied Spatial Data Analysis with R by Bivand, Roger S. and Pebesma, Edzer and Gomez-Rubio and Virgilio

See also

Examples

if (interactive()) {

if (requireNamespace("rworldmap", quietly = TRUE)) {
    world <- rworldmap::getMap(resolution = "coarse")
    p <- l_plot()
    lmap <- l_layer(p, world, asSingleLayer=TRUE)
    l_scaleto_world(p)
    attr(lmap,'hole')
    attr(lmap,'NAME')
}
}