layer a group node

Loon's displays that are based on Cartesian coordinates (i.e. scatterplot, histogram and graph display) allow for layering visual information including polygons, text and rectangles.

A group layer can contain other layers. If the group layer is invisible, then so are all its children.

Usage

l_layer_group(widget, label = "group", parent = "root", index = 0)

Arguments

widget

widget path name as a string

label

label used in the layers inspector

parent

group layer

index

of the newly added layer in its parent group

Value

layer object handle, layer id

Details

For more information run: l_help("learn_R_layer")

See also

Examples

if (interactive()){
p <- l_plot(x=c(1,10,1.5,7,4.3,9,5,2,8),
             y=c(1,10,7,3,4,3.3,8,3,4),
             title="Demo Layers")

id.g <- l_layer_group(p, "A Layer Group")
id.pts <- l_layer_points(p, x=c(3,6), y=c(4,7), color="red", parent=id.g)
l_scaleto_layer(p, id.pts)
l_configure(id.pts, x=c(-5,5,12), y=c(-2,-5,18), color="lightgray")

}