Layer a oval

Loon's displays that are based on Cartesian coordinates (i.e. scatterplot, histogram and graph display) allow for layering visual information including polygons, text and rectangles.

Usage

l_layer_oval(
  widget,
  x,
  y,
  color = "gray80",
  linecolor = "black",
  linewidth = 1,
  label = "oval",
  parent = "root",
  index = 0,
  ...
)

Arguments

widget

widget path name as a string

x

x coordinates

y

y coordinates

color

fill color, if empty string "", then the fill is transparant

linecolor

outline color

linewidth

linewidth of outline

label

label used in the layers inspector

parent

group layer

index

of the newly added layer in its parent group

...

additional state initialization arguments, see l_info_states

Value

layer object handle, layer id

Details

For more information run: l_help("learn_R_layer")

See also

Examples

if(interactive()){

p <- l_plot()
l <- l_layer_oval(p, c(1,5), c(2,12), color='steelblue')
l_configure(p, panX=0, panY=0, deltaX=20, deltaY=20)

}