2d navigation graph setup with with dynamic node fitering using a slider

Generic function to create a navigation graph environment where user can filter graph nodes using as slider to select 2d spaces based on 2d measures.

l_ng_ranges(measures, ...)

Arguments

measures
object with measures are stored

...
argument passed on to methods

Details

For more information run: l_help("learn_R_display_graph.html#l_ng_ranges")

See also