l_plot method for classical seasonal decomposition of a time series by moving averages (i.e. via decompose)

a generic function for creating a decomposed time seires plot based on objects of class decomposed.ts

# S3 method for decomposed.ts
l_plot(x, y = NULL, pcolor = NULL, size = 1,
  lcolor = NULL, linewidth = 1, xlabels = NULL, ylabels = NULL,
  title = NULL, tk_title = NULL, linkingGroup = NULL,
  showScales = TRUE, showGuides = TRUE, showLabels = TRUE, ...)

Arguments

x
a decomposed.ts object

y
NULL, ignored

pcolor
points colour of all time series. If NULL (the default) lcolor will be l_getOption("foreground").

size
points size of all time series. Default value is 1.

lcolor
line colour of all time series. If NULL (the default) lcolor will be l_getOption("foreground").

linewidth
line width of all time series (incl. original and decomposed components. Default is 1.

xlabels
the labels for the x axes: a length four character vector for each of the original time series, trend, seasonality and remainder. If NULL, xlabels is created as "time".

ylabels
the labels for the vertical axes: a length four character vector for each of the original time series, trend, seasonality and remainder. If NULL, ylabels is created based on the information provided.

title
an overall title for the entire display. If NULL (the default), the title will be "Seasonal Trend Analysis".

tk_title
provides an alternative window name to Tk's wm title. If NULL, stl will be used.

linkingGroup
name of linking group. If NULL, one is created from the data name and class associated with stlOrDecomposedTS.

showScales
a logical as to whether to display the scales on all axes, default is TRUE.

showGuides
a logical as to whether to display background guide lines on all plots, default is TRUE.

showLabels
a logical as to whether to display axes labels on all plots, default is TRUE.

...
keyword value pairs passed off to l_plot() which constructs each loon scatterplot component.

Value

A structure of class "l_ts" containing four loon plots each representing a part of the decomposition by name: "original", "trend", "seasonal", and "remainder".

Examples

# NOT RUN {
decompose <- decompose(co2)
p <- l_plot(decompose, title = "Atmospheric carbon dioxide over Mauna Loa")
# names of plots in the display
names(p)
# names of states associated with the seasonality plot
names(p$seasonal)
# which can be set
p$seasonal['color'] <- "steelblue"
# }