The l_plot method for density objects.

Creates an loon plot displaying contours of the supplied (kernel) density estimate

# S3 method for density
l_plot(x, y = NULL, xlabel = NULL, ylabel = NULL,
  title = NULL, linewidth = 2, color = NULL, size = 0.5, ...)

Arguments

x
a density object

y
NULL, ignored.

xlabel
the graphical parameter xlabel labelling the x axis of the plot. If NULL (the default), an xlabel is created based on the information available from the density objects.

ylabel
the graphical parameter ylabel labelling the y axis of the plot. If NULL (the default), an ylabel is created based on the combination of data name and "density".

title
providing a title for the plot. If NULL (the default), the title will be the call which produced the result.

linewidth
line width, default value is 2.

color
line colour. If NULL (the default) colour will be taken from l_getOption("foreground").

size
points size, default value is 0.5

...
named arguments being states passed to l_plot()

See also

Examples

ds <- density(faithful$eruptions)
p <- l_plot(ds, color = "steelblue")