Set loon's color mapping list to the colors from ColorBrewer

Loon's color list is used to map nominal values to colors. See the documentation for l_setColorList.

l_setColorList_ColorBrewer(palette = c("Set1", "Set2", "Set3", "Pastel1",
  "Pastel2", "Paired", "Dark2", "Accent"))

Arguments

palette
one of the following RColorBrewer palette name: Set1, Set2, Set3, Pastel1, Pastel2, Paired, Dark2, or Accent

Details

Only the following palettes in ColorBrewer are available: Set1, Set2, Set3, Pastel1, Pastel2, Paired, Dark2, and Accent. See the examples below.

See also

Examples

if (FALSE) {
if (requireNamespace("RColorBrewer", quietly = TRUE)) {
  RColorBrewer::display.brewer.all()
}
}

l_setColorList_ColorBrewer("Set1")
p <- l_plot(iris)