Open a browser with loon's R documentation webpage

l_web opens a browser with the relevant page on the official loon documentation website at http://great-northern-diver.github.io/loon/.

l_web(page = "index", directory = c("home", "reference", "articles"),
  ...)

Arguments

page
relative path to a page, the .html part may be omitted

directory
if "home" then page is ignored and the browser will open at the home page of the official loon documentation website at http://great-northern-diver.github.io/loon/. If page refers to a loon manual reference, then directory must be "reference"; if page refers to the name of a vignette file, then directory should be "articles"

...
arguments forwarded to browseURL, e.g. to specify a browser

See also

Examples

# NOT RUN {
l_web()
#
vignette("introduction", package = "loon")
# or
l_web(page = "introduction", directory = "articles")
#
help(l_hist)
l_web(page = "l_hist", directory = "reference")
# }