Loon's color generator for creating color palettes

Loon has a color sequence generator implemented creates a color palettes where the first m colors of a color palette of size m+1 are the same as the colors in a color palette of size m, for all positive natural numbers m. See the details in the l_setColorList documentation.

Usage

loon_palette(n)

Arguments

n

number of different colors in the palette

Value

vector with hex-encoded color values

See also

Examples

loon_palette(12)
#>  [1] "#68B2ED" "#CDA64D" "#EF8AC9" "#1EC28B" "#C698EE" "#92B74D" "#EE938C"
#>  [8] "#00C1C6" "#9FA5F3" "#B3AF40" "#F38CAC" "#00C3AA"