Closure of One Dimensional Measures

Function creates a 1d measures object that can be used with l_ng_plots and l_ng_ranges.

measures1d(data, ...)

Arguments

data
a data.frame with the data used to calculate the measures

...
named arguments, name is the function name and argument is the function to calculate the measure for each variable.

Value

a measures object

Details

For more information run: l_help("learn_R_display_graph.html#measures")

See also

Examples

m1 <- measures1d(oliveAcids, mean=mean, median=median,
     sd=sd, q1=function(x)as.vector(quantile(x, probs=0.25)),
     q3=function(x)as.vector(quantile(x, probs=0.75)))
     
m1

#> [1] "1d measure closure with measures: mean, median, sd, q1, q3 \n"

m1()

#>                   mean median        sd      q1      q3
#> palmitic    1231.74126 1201.0 168.59226 1095.00 1360.00
#> palmitoleic  126.09441  110.0  52.49436   87.75  169.25
#> stearic      228.86538  223.0  36.74494  205.00  249.00
#> oleic       7311.74825 7302.5 405.81022 7000.00 7680.00
#> linoleic     980.52797 1030.0 242.79922  770.75 1180.75
#> linolenic     31.88811   33.0  12.96870   26.00   40.25
#> arachidic     58.09790   61.0  22.03025   50.00   70.00
#> eicosenoic    16.28147   17.0  14.08330    2.00   28.00

m1(olive$palmitoleic>100)

#>                   mean median        sd      q1      q3
#> palmitic    1331.49693 1345.5 144.92782 1254.25 1419.00
#> palmitoleic  162.35890  163.0  39.13201  130.00  185.00
#> stearic      224.03374  220.0  33.35324  203.00  242.00
#> oleic       7058.02454 7025.0 282.79153 6883.00 7253.75
#> linoleic    1090.95399 1134.0 177.25662  950.00 1215.50
#> linolenic     33.86196   34.0  10.89852   29.00   40.75
#> arachidic     59.93252   61.0  17.65329   53.00   68.00
#> eicosenoic    21.59816   23.5  11.18355   16.00   29.00

m1('data')

#>     palmitic palmitoleic stearic oleic linoleic linolenic arachidic eicosenoic
#> 1       1075          75     226  7823      672        36        60         29
#> 2       1088          73     224  7709      781        31        61         29
#> 3        911          54     246  8113      549        31        63         29
#> 4        966          57     240  7952      619        50        78         35
#> 5       1051          67     259  7771      672        50        80         46
#> 6        911          49     268  7924      678        51        70         44
#> 7        922          66     264  7990      618        49        56         29
#> 8       1100          61     235  7728      734        39        64         35
#> 9       1082          60     239  7745      709        46        83         33
#> 10      1037          55     213  7944      633        26        52         30
#> 11      1051          35     219  7978      605        21        65         24
#> 12      1036          59     235  7868      661        30        62         44
#> 13      1074          70     214  7728      747        50        79         33
#> 14       875          52     243  8018      655        41        79         32
#> 15       952          49     254  7795      780        50        75         41
#> 16      1155          98     201  7606      816        32        60         29
#> 17       943          94     183  7840      788        42        75         31
#> 18      1278          69     205  7344      957        45        70         28
#> 19       961          70     195  7958      742        46        75         30
#> 20       952          77     258  7820      736        43        78         33
#> 21      1074          67     236  7692      716        56        83         45
#> 22       995          46     288  7806      679        56        86         40
#> 23      1056          53     247  7703      700        54        89         51
#> 24      1065          39     234  7876      703        42        74         26
#> 25      1065          45     245  7779      696        47        82         38
#> 26      1315         139     230  7299      832        42        60         32
#> 27      1321         136     217  7174      950        43        63         30
#> 28      1359         115     246  7234      874        45        63         18
#> 29      1378         111     272  7127      940        46        64         23
#> 30      1295         109     245  7253      903        43        62         38
#> 31      1275         121     215  7285      892        40        68         41
#> 32      1336         120     318  7083      915        50        70         38
#> 33      1309         122     241  7257      870        46        72         35
#> 34      1340         114     189  7337      820        48        72         21
#> 35      1299         116     253  7309      823        40        69         27
#> 36      1221         107     221  7441      798        54        70         28
#> 37      1245          72     283  7395      829        44        67         28
#> 38      1285         129     244  7323      819        57        65         36
#> 39      1248         107     313  7299      840        46        66         33
#> 40      1356         106     236  7209      866        48        75         36
#> 41      1260         102     228  7354      870        49        64         28
#> 42      1261         121     312  7238      877        47        65         25
#> 43      1304         124     279  7160      928        48        61         37
#> 44      1344         117     287  7129      897        51        65         41
#> 45      1323          96     300  7351      757        47        54         26
#> 46      1292         117     215  7351      839        48        61         32
#> 47      1254         118     244  7394      786        46        71         24
#> 48      1312         131     259  7167      939        41        69         20
#> 49      1213         109     301  7261      925        47        65         31
#> 50      1359          98     351  7262      780        41        56         16
#> 51      1266          97     263  7435      743        45        69         29
#> 52      1298          99     311  7311      787        45        67         23
#> 53      1272         116     279  7258      872        43        72         27
#> 54      1278          87     332  7379      771        44        53         24
#> 55      1184         112     311  7391      819        48        57         28
#> 56      1382         110     268  7241      828        39        60         30
#> 57      1183         146     292  7580      618        38        51         23
#> 58      1261         153     219  7355      818        52        70         26
#> 59      1198         136     239  7639      633        27        55         19
#> 60      1225         134     232  7658      616        36        49         26
#> 61      1339         166     208  7190      923        40        69         25
#> 62      1132         157     240  7641      638        45        60         31
#> 63      1381         183     245  7385      609        47        70         25
#> 64      1409         128     257  7257      759        43        57         16
#> 65      1306         127     250  7254      869        47        68         24
#> 66      1372         120     250  7355      702        44        68         28
#> 67      1336         113     242  7293      855        38        60         18
#> 68      1401         151     238  7164      857        45        72         36
#> 69      1390         119     234  7236      823        40        62         41
#> 70      1432         152     281  7029      949        39        55         25
#> 71      1412         124     298  7182      790        45        68         28
#> 72      1366         147     291  7197      783        51        70         34
#> 73      1383         118     273  7282      738        45        68         29
#> 74      1283         102     263  7400      763        54        65         28
#> 75      1296         136     260  7380      780        48        51         18
#> 76      1287         108     287  7343      826        44        44         23
#> 77      1351         159     296  7229      810        36        60         22
#> 78      1241          97     268  7499      709        52        69         36
#> 79      1267         101     300  7230      898        74        65         34
#> 80      1235         138     252  7322      861        54        66         36
#> 81      1255         103     223  7395      848        47        56         30
#> 82      1454         183     196  7057     1014        27        46         19
#> 83      1347         194     197  7277      895        25        46         15
#> 84      1364         204     225  6929     1084        21        50         14
#> 85      1410         199     216  7130      955        21        48         19
#> 86      1384         178     208  7105      999        29        67         26
#> 87      1412         185     217  6842     1203        34        72         32
#> 88      1410         232     280  6715     1233        32        60         24
#> 89      1509         209     257  6647     1240        42        62         30
#> 90      1317         197     256  7036     1067        40        60         22
#> 91      1286         192     203  7132     1053        38        65         28
#> 92      1273         191     202  6862     1303        43        70         28
#> 93      1463         183     183  6747     1307        36        60         24
#> 94      1399         187     191  6861     1233        38        60         17
#> 95      1413         193     208  6875     1202        30        60         18
#> 96      1369         206     203  6953     1168        35        50         16
#> 97      1488         172     170  6920     1144        37        54         14
#> 98      1323         160     205  6911     1298        24        50         17
#> 99      1311         166     170  6902     1312        41        69         28
#> 100     1286         163     183  7040     1230        29        57         12
#> 101     1380         173     188  7038     1139        31        44         14
#> 102     1394         164     223  7086     1042        24        43         23
#> 103     1324         174     198  6863     1289        36        70         21
#> 104     1290         157     192  7000     1263        26        51         19
#> 105     1361         163     196  6888     1273        37        58         24
#> 106     1387         182     242  6913     1101        44        68         30
#> 107     1369         180     181  7000     1130        39        45         24
#> 108     1303         165     175  7025     1243        31        41         16
#> 109     1346         160     169  7072     1151        39        48         15
#> 110     1369         171     184  6937     1246        30        48         15
#> 111     1305         172     169  7004     1260        28        50         11
#> 112     1351         179     186  6935     1243        36        50         19
#> 113     1283         151     182  7000     1271        40        52         21
#> 114     1449         175     198  6883     1162        40        70         22
#> 115     1310         180     183  7054     1202        26        32         12
#> 116     1360         163     176  6901     1280        28        65         27
#> 117     1300         187     196  6920     1253        41        76         25
#> 118     1368         171     218  7010     1057        41        54         26
#> 119     1207         151     156  7159     1234        27        51         14
#> 120     1348         154     183  6917     1277        48        56         16
#> 121     1334         186     229  7261      827        34        56         20
#> 122     1301         156     207  7003     1229        41        48         14
#> 123     1226         181     213  6961     1230        47        74         26
#> 124     1201         168     190  7100     1216        43        64         16
#> 125     1297         153     177  7004     1260        35        60         16
#> 126     1248         163     158  7103     1222        31        60         14
#> 127     1335         159     197  6974     1220        36        60         17
#> 128     1219         167     171  7087     1254        35        50         16
#> 129     1318         179     177  7030     1194        35        42         25
#> 130     1264         167     166  7130     1187        22        52         12
#> 131     1201         175     201  7129     1193        36        49         15
#> 132     1252         180     181  7055     1214        31        59         38
#> 133     1273         182     209  6965     1191        43        74         23
#> 134     1351         179     170  7034     1154        35        66         10
#> 135     1336         155     212  7103     1086        33        55         20
#> 136     1499         201     182  6803     1204        30        56         24
#> 137     1425         198     193  7032     1041        31        52         17
#> 138     1358         204     227  6962     1109        41        65         34
#> 139     1346         181     257  7147      933        40        60         36
#> 140     1392         186     256  6732     1278        53        64         29
#> 141     1311         166     222  7006     1147        41        80         27
#> 142     1314         171     229  6923     1198        47        76         42
#> 143     1409         200     207  6842     1224        31        60         27
#> 144     1342         174     221  6993     1147        36        64         23
#> 145     1387         182     206  7100     1020        34        54         17
#> 146     1413         202     205  6920     1165        36        46         13
#> 147     1430         209     225  6800     1200        32        59         27
#> 148     1336         185     223  6956     1155        56        73         16
#> 149     1372         200     200  6916     1189        33        50         22
#> 150     1330         157     228  7055     1108        42        55         25
#> 151     1412         207     208  6822     1239        36        51         28
#> 152     1321         209     217  6948     1178        42        62         23
#> 153     1401         200     217  6980     1073        40        68         21
#> 154     1401         214     217  6734     1293        44        69         27
#> 155     1457         168     242  6724     1266        54        59         30
#> 156     1451         199     221  6835     1177        37        51         29
#> 157     1438         206     248  6806     1183        34        57         28
#> 158     1462         204     237  6644     1309        42        54         28
#> 159     1529         215     203  6602     1310        45        69         27
#> 160     1510         189     245  6752     1188        36        52         28
#> 161     1437         222     184  6803     1240        43        56         16
#> 162     1327         129     247  7024     1157        38        56         22
#> 163     1438         172     252  6630     1380        40        64         24
#> 164     1447         176     189  6849     1180        42        64         26
#> 165     1355         144     214  6972     1198        33        60         24
#> 166     1369         156     241  6890     1209        42        63         30
#> 167     1471         188     276  6697     1269        34        51         16
#> 168     1456         179     240  6738     1267        41        65         14
#> 169     1314         140     207  7020     1220        28        59         12
#> 170     1408         176     192  6909     1195        45        50         25
#> 171     1397         172     191  7107     1018        36        50         29
#> 172     1413         191     186  6937     1180        31        46         13
#> 173     1539         194     213  6764     1178        38        58         16
#> 174     1304         159     234  7019     1174        38        53         19
#> 175     1341         160     231  7033     1069        40        67         33
#> 176     1508         208     249  6641     1311        25        43         20
#> 177     1515         226     257  6595     1287        41        63         16
#> 178     1262         165     235  7120     1113        32        51         21
#> 179     1307         197     238  7003     1144        37        50         24
#> 180     1294         159     253  7009     1190        30        52         13
#> 181     1460         187     215  6843     1172        35        56         32
#> 182     1476         187     203  6837     1197        36        48         22
#> 183     1482         178     197  6814     1201        40        64         24
#> 184     1388         176     185  7008     1111        48        53         31
#> 185     1367         172     235  7066     1054        35        45         26
#> 186     1272         207     205  7152     1098        37        52         22
#> 187     1323         157     234  7132     1022        38        58         31
#> 188     1206         218     242  7193     1002        37        54         25
#> 189     1383         157     217  7018     1090        40        60         37
#> 190     1521         190     238  6956      986        36        50         23
#> 191     1350         168     227  6986     1165        29        58         17
#> 192     1422         181     218  6813     1230        30        59         21
#> 193     1298         166     224  6986     1162        34        65         31
#> 194     1447         236     245  6607     1336        33        51         21
#> 195     1347         197     211  6795     1300        32        59         34
#> 196     1339         170     253  6989     1110        29        63         23
#> 197     1388         183     216  6867     1208        28        61         21
#> 198     1527         260     232  6488     1370        31        45         20
#> 199     1495         237     236  6571     1318        32        58         26
#> 200     1487         246     251  6504     1390        29        53         19
#> 201     1399         180     232  6855     1190        32        66         22
#> 202     1489         215     242  6777     1145        30        60         22
#> 203     1339         166     226  6928     1198        30        60         23
#> 204     1482         246     238  6444     1462        27        50         20
#> 205     1434         172     255  6646     1354        27        59         25
#> 206     1347         156     214  6850     1313        25        48         19
#> 207     1340         158     233  6848     1272        32        63         25
#> 208     1453         180     244  6752     1238        34        54         23
#> 209     1306         149     226  7082     1097        33        61         24
#> 210     1349         161     217  6997     1138        31        62         23
#> 211     1254         151     205  7319      947        28        54         23
#> 212     1168         144     220  7230     1109        31        52         28
#> 213     1346         167     224  6959     1111        30        49         23
#> 214     1390         184     212  6898     1189        29        44         19
#> 215     1283         149     224  7077     1104        30        57         32
#> 216     1214         137     232  7269     1005        32        55         23
#> 217     1491         227     205  6941      988        33        68         34
#> 218     1479         218     207  7039      887        36        65         36
#> 219     1445         174     228  6875     1123        29        69         31
#> 220     1439         183     218  6775     1226        32        66         29
#> 221     1387         154     204  6991     1090        34        74         32
#> 222     1426         169     192  7025     1043        31        64         27
#> 223     1451         200     208  6980     1006        30        62         31
#> 224     1493         204     188  6913     1044        32        61         35
#> 225     1419         192     207  6996     1014        36        70         36
#> 226     1342         177     199  7172      952        34        65         33
#> 227     1349         152     236  7145      949        35        75         29
#> 228     1440         196     208  6938     1070        32        61         26
#> 229     1460         215     197  6918     1081        28        55         23
#> 230     1249         133     205  7417      827        33        72         33
#> 231     1348         159     238  7017     1081        31        67         25
#> 232     1341         155     244  6958     1144        32        68         26
#> 233     1398         149     204  7182      907        29        76         30
#> 234     1454         200     199  6910     1090        30        62         25
#> 235     1334         153     219  6928     1214        33        66         24
#> 236     1438         204     189  7107      910        33        63         27
#> 237     1303         138     212  7170     1016        34        69         25
#> 238     1323         147     210  7108     1070        33        61         20
#> 239     1417         169     207  6875     1184        34        57         27
#> 240     1360         167     225  6883     1220        31        55         27
#> 241     1420         179     214  6923     1121        33        56         27
#> 242     1472         218     214  6724     1238        29        53         23
#> 243     1368         174     205  7042     1066        31        57         26
#> 244     1367         173     228  6948     1141        32        53         24
#> 245     1403         173     209  6843     1210        33        63         33
#> 246     1413         197     206  6737     1387        34        60         31
#> 247     1201         138     207  7011     1269        37        64         35
#> 248     1359         180     207  6895     1203        33        61         30
#> 249     1518         198     225  6681     1243        29        57         24
#> 250     1434         185     189  6771     1269        30        62         25
#> 251     1367         162     179  6772     1368        33        64         27
#> 252     1461         181     197  6783     1246        26        57         23
#> 253     1368         161     198  7030     1095        33        59         31
#> 254     1419         159     215  6862     1193        35        60         31
#> 255     1514         162     298  6725     1119        45        93         30
#> 256     1328         171     253  6987     1030        38        83         39
#> 257     1469         160     337  6675     1127        44        94         36
#> 258     1222         133     227  7425      824        36        69         35
#> 259     1639         172     331  6510     1124        46        91         32
#> 260     1345         133     272  6801     1194        48        83         37
#> 261     1339         170     275  6838     1060        46        88         43
#> 262     1194         135     263  7277      889        44        95         41
#> 263     1112          68     375  7770      448        52        69         45
#> 264     1222          70     329  7605      566        48        67         43
#> 265     1136          72     341  7616      661        49        65         32
#> 266      926          41     277  7815      784        45        65         25
#> 267     1105          69     373  7714      532        51        68         37
#> 268     1109          79     305  7576      763        45        64         36
#> 269     1284          93     265  7235      893        43        77         46
#> 270     1120          69     277  7416      946        42        59         36
#> 271      916          52     281  7870      694        42        64         58
#> 272      905          49     288  7747      812        49        71         56
#> 273     1206          55     287  7329      935        44        74         42
#> 274     1457         182     267  7020      863        41        84         37
#> 275     1327         140     193  7328      823        36        87         35
#> 276     1303         100     251  7045     1049        40        86         40
#> 277     1444         175     259  6876     1027        34        78         32
#> 278     1505         243     226  6962      858        30        72         27
#> 279     1429         162     223  6917     1041        37        77         40
#> 280     1491         162     211  6994      928        37        97         38
#> 281     1393         128     211  7189      870        38        93         40
#> 282     1404         134     210  7110      923        40       101         43
#> 283     1222         130     214  7374      856        38        89         45
#> 284     1153          74     316  7593      705        42        64         32
#> 285     1169          76     307  7553      728        43        69         32
#> 286     1369         104     237  7375      775        39        70         15
#> 287      993          58     267  7743      773        41        62         44
#> 288      980          53     254  7719      815        44        69         47
#> 289      967          55     273  7692      833        45        63         47
#> 290     1128          73     354  7527      728        44        76         38
#> 291     1188          85     273  7445      814        44        73         42
#> 292     1257          95     247  7405      812        43        70         35
#> 293     1262          88     301  7471      704        43        71         31
#> 294     1283         153     196  7107     1115        37        60         28
#> 295     1263         155     199  7140     1148        31        42         18
#> 296     1369         158     215  7160      958        38        69         32
#> 297     1353         172     175  6965     1212        28        75         19
#> 298     1187         139     185  7427      952        29        56         22
#> 299     1732         231     156  6437     1313        45        62         23
#> 300     1620         255     166  6628     1212        29        62         27
#> 301     1543         172     193  6740     1157        52        87         34
#> 302     1498         170     195  6804     1206        35        66         23
#> 303     1399         169     171  7011     1100        36        72         16
#> 304     1293         156     191  7101     1111        32        60         31
#> 305     1420         175     152  7004     1149        27        50         20
#> 306     1721         238     255  6300     1350        35        70         28
#> 307     1742         221     156  6415     1315        43        82         23
#> 308     1391         187     189  6975     1062        52        70         45
#> 309     1517         206     249  6680     1205        33        80         27
#> 310     1269         157     193  7140     1148        31        40         18
#> 311     1577         204     208  6732     1183        20        52         20
#> 312     1590         241     195  6705     1149        27        68         21
#> 313     1621         280     197  6608     1179        28        58         27
#> 314     1753         275     236  6367     1214        23        61         27
#> 315     1679         260     177  6568     1191        30        59         33
#> 316     1419         203     176  6973     1083        38        78         27
#> 317     1693         236     174  6499     1204        51       102         37
#> 318     1692         270     234  6499     1196        31        59         15
#> 319     1638         252     215  6570     1199        39        53         29
#> 320     1497         247     219  6621     1270        36        73         32
#> 321     1442         222     194  6677     1314        36        72         38
#> 322     1680         270     170  6440     1310        31        62         28
#> 323     1463         164     185  6909     1154        49        58         17
#> 324     1129         120     222  7272     1112        43        98          2
#> 325     1042         135     210  7376     1116        35        90          3
#> 326     1103          96     210  7380     1085        32        94          3
#> 327     1118          97     221  7279     1154        35        94          2
#> 328     1052          95     215  7388     1126        31        92          1
#> 329     1116         102     231  7290     1168        26        66          1
#> 330     1108         132     231  7319     1101        20        66          2
#> 331     1129         108     212  7386     1074        28        62          3
#> 332     1085          91     223  7384     1126        28        62          3
#> 333     1104         103     233  7322     1147        27        61          2
#> 334     1098          88     212  7338     1140        28        67          1
#> 335     1135          98     251  7120     1314        20        61          2
#> 336     1158         108     245  7065     1326        22        75          1
#> 337     1133         110     241  7080     1342        21        68          3
#> 338     1095         125     250  7120     1305        21        83          1
#> 339     1201          87     238  6990     1383        25        75          3
#> 340     1213         112     245  7007     1335        22        65          3
#> 341     1108          92     231  7367     1110        29        62          3
#> 342     1075         103     207  7413     1096        32        68          2
#> 343     1059          96     228  7386     1128        25        72          2
#> 344     1176          92     207  7347     1057        35        82          1
#> 345     1159          98     213  7320     1108        38        64          1
#> 346     1132          80     201  7398     1095        27        67          2
#> 347     1107          75     220  7399     1096        29        90          1
#> 348     1092         104     234  7355     1126        28        58          2
#> 349     1119          81     219  7409     1057        33        81          2
#> 350     1106          93     212  7381     1104        35        68          1
#> 351     1047         101     238  7385     1120        28        89          1
#> 352     1165          99     214  7331     1101        22        67          3
#> 353     1158          84     201  7327     1123        29        77          2
#> 354     1095          88     203  7415     1093        37        78          1
#> 355     1176          75     205  7396     1107        33        74          2
#> 356     1103         109     220  7335     1140        28        59          2
#> 357     1112          92     209  7356     1125        32        73          2
#> 358     1091          93     222  7377     1113        20        53          2
#> 359     1080          98     219  7371     1125        33        78          1
#> 360     1051         108     227  7403     1114        30        66          3
#> 361     1096          84     211  7415     1091        30        71          2
#> 362     1142          97     225  7341     1101        28        65          1
#> 363     1047          96     236  7399     1107        32        80          3
#> 364     1114          86     210  7359     1116        31        83          2
#> 365     1140          93     241  7324     1098        23        74          1
#> 366     1075          91     200  7410     1107        36        80          1
#> 367     1092         106     219  7427     1125        33        77          1
#> 368     1076          95     204  7408     1130        27        79          2
#> 369     1178          89     201  7381     1099        34        87          2
#> 370     1095         104     223  7367     1111        43        56          2
#> 371     1166          97     272  6971     1390        20        83          3
#> 372     1154         119     257  7130     1253        22        61          1
#> 373     1177         111     241  6882     1470        22        95          2
#> 374     1160          96     240  7043     1357        24        79          2
#> 375     1122         104     241  7145     1313        15        58          1
#> 376     1132          99     257  7065     1362        24        90          3
#> 377     1096         100     260  7162     1282        25        74          2
#> 378     1131          87     233  7144     1307        25        72          3
#> 379     1184         105     258  7020     1340        26        66          2
#> 380     1135          94     235  7123     1320        24        67          2
#> 381     1084          96     240  7164     1330        28        57          1
#> 382     1086         127     252  7159     1285        28        62          2
#> 383     1140          95     258  7085     1347        23        71          3
#> 384     1138         101     254  7103     1310        25        68          1
#> 385     1159         110     261  7068     1297        27        77          2
#> 386     1051          78     211  7421     1146        30        82          2
#> 387     1048          79     213  7439     1130        28        61          2
#> 388     1061          86     220  7421     1102        29        79          3
#> 389     1105          88     210  7353     1142        28        72          1
#> 390     1145          35     237  7208     1118        20        46          2
#> 391     1049          96     219  7303     1168        22        47          2
#> 392     1105         120     218  7302     1158        23        45          3
#> 393     1030          84     214  7403     1177        21        70          1
#> 394     1070          98     215  7280     1240        28        68          3
#> 395     1103          81     208  7310     1177        30        90          3
#> 396     1040         101     205  7368     1176        25        85          3
#> 397     1100          95     210  7320     1113        22        72          3
#> 398     1118          85     199  7415     1060        36        86          3
#> 399     1065          98     230  7345     1163        24        74          1
#> 400     1131          78     221  7358     1120        22        69          2
#> 401     1080         120     218  7296     1145        35       105          2
#> 402     1075          86     231  7403     1109        22        73          3
#> 403     1040         103     228  7364     1173        25        66          2
#> 404     1128          82     203  7320     1148        30        88          1
#> 405     1060         111     231  7363     1149        20        65          1
#> 406     1103          78     220  7365     1149        20        65          2
#> 407     1110          91     201  7318     1185        24        74          2
#> 408     1091         108     218  7383     1183        28        88          3
#> 409     1094          96     220  7341     1127        26        96          2
#> 410     1131          87     208  7170     1308        28        57          2
#> 411     1175         108     214  7076     1307        33        85          2
#> 412     1076          77     202  7243     1305        29        67          1
#> 413     1120          90     240  7068     1383        23        75          1
#> 414     1152         111     238  7080     1372        25        81          2
#> 415     1141          95     250  7035     1388        22        68          2
#> 416     1098         103     267  7135     1301        24        76          2
#> 417     1126         100     236  7062     1380        26        69          1
#> 418     1087          89     243  7200     1302        18        60          1
#> 419     1115          96     236  7085     1372        20        75          2
#> 420     1178          92     241  7006     1376        22        84          1
#> 421     1162         106     242  7025     1368        25        71          2
#> 422     1085          70     180  7955      605        20        50          1
#> 423     1085          70     185  7955      600        25        55          1
#> 424     1090          60     190  7950      600        28        47          2
#> 425     1080          65     189  7960      602        35        20          1
#> 426     1090          60     195  7955      600        28        42          2
#> 427     1105          55     200  7900      600        37        55          2
#> 428     1060          75     175  7975      610        20        55          2
#> 429     1050          70     170  7977      605        28        65          1
#> 430     1100          55     198  7905      600        35        50          3
#> 431     1065          65     178  7965      605        22        65          2
#> 432     1085          60     188  7955      602        30        50          2
#> 433     1080          65     180  7960      605        25        55          1
#> 434     1085          60     190  7955      602        30        53          1
#> 435     1075          68     195  7960      602        20        40          3
#> 436     1090          58     192  7950      600        35        40          3
#> 437     1095          60     198  7945      600        38        34          2
#> 438     1090          58     195  7950      600        30        42          2
#> 439     1095          58     198  7950      602        35        32          1
#> 440     1090          58     195  7940      600        35        42          2
#> 441     1095          58     198  7945      600        35        34          1
#> 442     1095          55     200  7940      600        35        45          3
#> 443     1080          70     188  7965      608        28        36          3
#> 444     1090          60     195  7950      600        32        38          2
#> 445     1105          55     200  7900      595        39        56          1
#> 446     1110          50     205  7900      595        40        52          1
#> 447     1075          70     198  7978      608        28        33          2
#> 448     1075          65     185  7980      608        35        42          3
#> 449     1065          75     180  7975      610        25        50          3
#> 450     1070          75     188  7980      602        22        45          2
#> 451     1070          75     188  7980      602        22        45          1
#> 452     1100          70     200  7910      610        39        44          1
#> 453     1075          70     185  7960      610        22        58          2
#> 454     1050          78     175  7990      610        18        59          3
#> 455     1090          60     198  7945      600        32        35          2
#> 456     1050          78     188  7990      608        28        23          3
#> 457     1075          70     190  7975      605        28        27          3
#> 458     1098          54     202  7945      595        42        32          2
#> 459     1105          15     198  8005      575        52        20          2
#> 460     1110          75     220  7915      510        55        65          2
#> 461     1058          50     178  7988      626        40        55          3
#> 462     1115          30     225  7955      600        55        15          2
#> 463     1105          30     198  7995      570        52        20          3
#> 464     1072          49     178  7980      615        48        48          2
#> 465     1110          15     210  7990      570        50        20          2
#> 466     1110          80     215  7910      525        50        60          1
#> 467     1055          60     175  7985      620        45        50          1
#> 468     1100          80     215  7930      535        45        60          2
#> 469     1105          55     205  7965      600        25        20          2
#> 470     1095          50     210  7948      600        25        35          2
#> 471     1110          50     220  7950      600        52        10          2
#> 472     1092          37     210  7955      600        40        40          3
#> 473     1290          60     260  7550      670        70       100          2
#> 474     1170          80     230  7690      720        40        70          1
#> 475     1100          90     250  7680      760        30        80          2
#> 476     1120          70     240  7720      730        40        80          2
#> 477     1160          70     250  7650      750        30        90          1
#> 478     1200          50     210  7770      690        20        50          3
#> 479     1140          50     200  7990      580        10        20          1
#> 480     1220          80     240  7610      760        30        60          2
#> 481     1180          90     250  7520      800        50       100          2
#> 482     1210          70     250  7560      780        40        90          2
#> 483     1220          80     220  7540      770        60       100          2
#> 484     1180         100     190  7520      820        50       100          1
#> 485     1160          90     220  7580      790        40        90          1
#> 486     1130         100     240  7620      780        30        90          1
#> 487     1080         100     260  7710      750        20        70          2
#> 488     1090          90     280  7730      720        50       100          1
#> 489     1020         100     270  7770      710        40        90          1
#> 490     1090          90     250  7680      760        60        80          1
#> 491     1120         100     260  7720      680        30        80          2
#> 492     1080          80     240  7830      670        30        70          2
#> 493     1160          70     230  7860      640        10        20          1
#> 494     1100          80     240  7820      670        20        70          2
#> 495     1050         100     250  7930      630        10        30          3
#> 496     1090          90     270  7780      690        30        50          3
#> 497     1120          80     260  7750      680        30        80          3
#> 498     1120         100     250  7680      730        40        70          2
#> 499     1190          90     230  7670      710        30        80          2
#> 500     1170         110     250  7620      740        20        90          1
#> 501     1120         100     230  7720      730        20        70          1
#> 502     1190          80     270  7690      720        10        40          2
#> 503     1400          90     270  7420      800         0        20          2
#> 504     1350          80     250  7520      760        10        30          1
#> 505     1090          60     220  7890      670        10        60          2
#> 506     1150          90     230  7790      650        30        60          1
#> 507     1240          90     220  7820      590        10        30          1
#> 508     1220         100     240  7890      530         0        10          2
#> 509     1180          80     250  7870      580        10        30          2
#> 510     1170         110     240  7730      630        30        90          1
#> 511     1170         100     280  7710      640        20        70          3
#> 512     1180          80     220  7790      680        10        40          1
#> 513     1200          90     240  7820      590        10        50          2
#> 514     1140          90     240  7880      570        20        60          3
#> 515     1160          70     210  7870      580        30        80          3
#> 516     1130          80     250  7780      650        40        60          3
#> 517     1150          80     240  7800      630        30        70          2
#> 518     1110          70     240  7820      670        20        70          3
#> 519     1150          70     220  7850      620        20        40          2
#> 520     1180          80     240  7760      670        20        50          2
#> 521     1020          80     250  7920      680        10        30          3
#> 522      610          80     230  8410      650         0        20          3
#> 523     1190         150     290  7340     1020         0        10          2
#> 524     1110         130     210  7550     1000         0         0          1
#> 525     1020         100     220  7530     1030         0         0          3
#> 526     1070         120     210  7600      990         0        10          3
#> 527     1010          90     350  7480     1050        10        10          1
#> 528     1060         140     240  7680      830        10        40          2
#> 529     1060         140     270  7620      880        10        20          1
#> 530     1030         100     230  7740      900         0         0          2
#> 531     1120         130     250  7530      970         0         0          3
#> 532     1030         110     220  7760      980         0         0          2
#> 533     1070         100     230  7600      990        10         0          1
#> 534     1140         180     220  7610      850        10        10          2
#> 535     1090         180     230  7590      860        10        40          2
#> 536      980         110     300  7720      910        10         0          3
#> 537      980          90     330  7540     1040         0         0          2
#> 538      960          90     200  7810      940         0         0          2
#> 539      990          90     210  7780      930         0         0          2
#> 540     1060         120     210  7600     1010         0         0          1
#> 541     1240         150     250  7610      730        10        10          1
#> 542     1060          90     310  7850      690         0         0          2
#> 543     1020         100     290  7620      960         0        10          2
#> 544      970          90     220  7700     1020         0         0          3
#> 545     1180         130     220  7450     1010         0        10          2
#> 546     1060         140     240  7690      850        10        10          1
#> 547      990         100     250  7630     1030         0         0          3
#> 548     1010          90     350  7630      940        10         0          3
#> 549     1040          90     250  7780      820        10        10          1
#> 550     1040          90     250  7810      810        10        10          2
#> 551     1020          90     350  7620      920        10         0          3
#> 552     1020          90     260  7620     1010         0         0          3
#> 553     1010          90     350  7610      930        10         0          3
#> 554      920         110     340  7720      910         0         0          3
#> 555     1030         100     250  7710      900         0        10          2
#> 556      960          90     300  7820      830         0         0          3
#> 557     1030         110     210  7810      840         0         0          1
#> 558     1010         100     240  7710      910        10        20          2
#> 559     1020          90     240  7800      850         0         0          2
#> 560     1120          90     300  7650      830         0        10          1
#> 561     1090          90     290  7710      800        10         0          2
#> 562     1100         120     280  7630      770        10        10          2
#> 563     1090          80     240  7820      760        10         0          2
#> 564     1150          90     250  7720      810         0        10          3
#> 565     1110          90     230  7810      750         0        10          2
#> 566     1010         110     210  7720      950         0         0          1
#> 567     1070         100     220  7730      870        10        10          2
#> 568     1280         110     290  7490      790        10        10          2
#> 569     1060         100     270  7740      810        10        10          3
#> 570     1010          90     210  7720      970         0         0          2
#> 571      990         120     250  7750      870        10        10          2
#> 572      960          80     240  7950      740        10        20          2

m1('measures')

#> [1] "mean"   "median" "sd"     "q1"     "q3"