Canadian Visible Minority Data 2006

Population census count of various named visible minority groups in each of 33 major census metropolitan areas of Canada in 2006.

These data are from the 2006 Canadian census, publicly available from Statistics Canada.

Usage

minority

Format

A data frame with 33 rows and 18 variates

Arab

Number identifying as `Arab`.

Black

Number identifying as `Black`.

Chinese

Number identifying as `Chinese`.

Filipino

Number identifying as `Filipino`.

Japanese

Number identifying as `Japanese`.

Korean

Number identifying as `Korean`.

Latin.American

Number identifying as `Latin American`.

Multiple.visible.minority

Number identifying as being a member of more than one visible minority.

South.Asian

Number identifying as `South Asian`.

Southeast.Asian

Number identifying as `Southeast Asian`.

Total.population

Total population of the metropolitan census area.

Visible.minority.not.included.elsewhere

Number identifying as a member of a visible minority that was not included elsewhere.

Visible.minority.population

Total number identifying as a member of some visible minority.

West.Asian

Number identifying as `West Asian`.

lat, long

Latitude and longitude (in degrees) of the metropolitan census area.

googleLat, googleLong

Latitude and longitude in degrees determined using the Google Maps Geocoding API.

rownames(minority) are the names of the metropolitan areas or cities.

Source

Statistics Canada