Closure of Two Dimensional Scagnostic Measures

Function creates a 2d measures object that can be used with l_ng_plots and l_ng_ranges.

scagnostics2d(
  data,
  scagnostics = c("Clumpy", "Monotonic", "Convex", "Stringy", "Skinny", "Outlying",
    "Sparse", "Striated", "Skewed"),
  separator = ":"
)

Arguments

data
a data.frame with the data used to calculate the measures

scagnostics
vector with valid scanostics meausure names, i.e "Clumpy", "Monotonic", "Convex", "Stringy", "Skinny", "Outlying", "Sparse", "Striated", "Skewed". Also the prefix "Not" can be added to each measure which equals 1-measure.

separator
string the separates variable names in 2d graph nodes

Value

a measures object

Details

For more information run: l_help("learn_R_display_graph.html#measures")

See also

Examples

if (FALSE) {
m <- scagnostics2d(oliveAcids, separator='**')
m
m()
m(olive$palmitoleic > 80)
m('data')
m('grid')
m('measures')
}