Layout as a grid

Layout as a grid

facet_grid_layout(
  plots,
  subtitles,
  by = NULL,
  prop = 10,
  parent = NULL,
  title = "",
  xlabel = "",
  ylabel = "",
  labelLocation = c("top", "right"),
  byrow = FALSE,
  swapAxes = FALSE,
  labelBackground = l_getOption("facetLabelBackground"),
  labelForeground = l_getOption("foreground"),
  labelBorderwidth = 2,
  labelRelief = "ridge",
  plotWidth = 200,
  plotHeight = 200,
  sep = "*",
  maxCharInOneRow = 15,
  new.toplevel = TRUE,
  ...
)

Arguments

plots
A list of loon plots

subtitles
The subtitles of the layout. It is a list and the length is equal to the number of by variables. Each element in a list is the unique values of such by variable.

by
an object of class "formula" (or one that can be coerced to that class): a symbolic description of the plots separated by

prop
The proportion of the label height and widget height

parent
a valid Tk parent widget path. When the parent widget is specified (i.e. not NULL) then the plot widget needs to be placed using some geometry manager like tkpack or tkplace in order to be displayed. See the examples below.

title
The title of the widget

xlabel
The xlabel of the widget

ylabel
The ylabel of the widget

labelLocation
Labels location.

Length two vector for layout grid. The first one is used to determine the position of column labels ('top' or 'bottom'). The second one is used to determine the position of row labels ('right' or 'left').

Length one vector for layout wrap, 'top' or 'bottom'.

byrow
Place widget by row or by column

swapAxes
swap axes, TRUE or FALSE

labelBackground
Label background color

labelForeground
Label foreground color

labelBorderwidth
Label border width

labelRelief
Label relief

plotWidth
default plot width (in pixel)

plotHeight
default plot height (in pixel)

sep
The character string to separate or combine a vector

maxCharInOneRow
Max char in one row. If it exceeds the max, then a char will be displayed into two rows

new.toplevel
determine whether the parent is a new top level. If it is not a new window, the widgets will not be packed

...
named arguments to modify plot states. See l_info_states of any instantiated l_plot for examples of names and values.