Layout Facets in a multiple panels

It takes a loon widget and forms a matrix of loon widget facets.

l_facet(widget, by, layout = c("grid", "wrap", "separate"), ...)

# S3 method for loon
l_facet(
  widget,
  by,
  layout = c("grid", "wrap", "separate"),
  connectedScales = c("cross", "row", "column", "both", "x", "y", "none"),
  linkingGroup,
  nrow = NULL,
  ncol = NULL,
  inheritLayers = TRUE,
  labelLocation = c("top", "right"),
  labelBackground = "gray80",
  labelForeground = "black",
  labelBorderwidth = 2,
  labelRelief = c("groove", "flat", "raised", "sunken", "ridge", "solid"),
  parent = NULL,
  ...
)

# S3 method for l_serialaxes
l_facet(
  widget,
  by,
  layout = c("grid", "wrap", "separate"),
  linkingGroup,
  nrow = NULL,
  ncol = NULL,
  labelLocation = c("top", "right"),
  labelBackground = "gray80",
  labelForeground = "black",
  labelBorderwidth = 2,
  labelRelief = c("groove", "flat", "raised", "sunken", "ridge", "solid"),
  parent = NULL,
  ...
)

Arguments

widget
A loon widget

by
Layout categories which can be loon n dimensional states, e.g. "color", "size", etc or a length n object. It takes a vector (single facet category), a data frame or a list (multiple facet categories).

layout
layout facets as 'grid', 'wrap' or 'separate'

...
named arguments to modify the `loon` widget states

connectedScales
Determines how the scales of the facets are to be connected depending on which layout is used. For each value of layout, the scales are connected as follows:

layout = "wrap:} Across all facets, when \code{connectedScales} is \itemize{ \item{\code{"x"}, then only the "x" scales are connected} \item{\code{"y"}, then only the "y" scales are connected} \item{\code{"both"}, both "x" and "y" scales are connected} \item{\code{"none"}, neither "x" nor "y" scales are connected.} For any other value, only the "y" scale is connected. } } \item{\code{layout = "grid: Across all facets, when connectedScales is

"cross", then only the scales in the same row and the same column are connected

"row", then both "x" and "y" scales of facets in the same row are connected

"column", then both "x" and "y" scales of facets in the same column are connected

"x", then all of the "x" scales are connected (regardless of column)

"y", then all of the "y" scales are connected (regardless of row)

"both", both "x" and "y" scales are connected in all facets

"none", neither "x" nor "y" scales are connected in any facets.

linkingGroup
A linkingGroup for widgets. If missing, default would be a paste of "layout" and the current tk path number.

nrow
The number of layout rows

ncol
The number of layout columns

inheritLayers
Logical value. Should widget layers be inherited into layout panels?

labelLocation
Labels location.

Length two vector for layout grid. The first one is used to determine the position of column labels ('top' or 'bottom'). The second one is used to determine the position of row labels ('right' or 'left').

Length one vector for layout wrap, 'top' or 'bottom'.

labelBackground
Label background color

labelForeground
Label foreground color

labelBorderwidth
Label border width

labelRelief
Label relief

parent
parent widget path

Value

an `l_facet` object (an `l_compound` object), being a list with named elements, each representing a separate interactive plot. The names of the plots should be self explanatory and a list of all plots can be accessed from the `l_facet` object via `l_getPlots()`.

Examples

library(maps)
p <- with(quakes, l_plot(long, lat, linkingGroup = "quakes"))
p["color"][quakes$mag < 5 & quakes$mag >= 4] <- "lightgreen"
p["color"][quakes$mag < 6 & quakes$mag >= 5] <- "lightblue"
p["color"][quakes$mag >= 6] <- "firebrick"
# A Fiji map
NZFijiMap <- map("world2", regions = c("New Zealand", "Fiji"), plot = FALSE)
l_layer(p, NZFijiMap,
        label = "New Zealand and Fiji",
        color = "forestgreen",
        index = "end")

#> loon layer "New Zealand and Fiji" of type polygons of plot .l16.plot
#> [1] "layer0"

fp <- l_facet(p, by = "color", layout = "grid",
              linkingGroup = "quakes")

# serialaxes facets
s <- l_serialaxes(iris, color = iris$Species,
                  scaling = "observation")
fs <- l_facet(s, layout = "wrap", by = iris$Species)

#> [1] "linkingGroup: facet20"

# The linkingGroup can be printed or accessed by
l_configure(s, linkingGroup = fs[[1]]['linkingGroup'], sync = "pull")