Create an interactive histogram

l_hist is a generic function for creating an interactive histogram display that can be linked with loon's other displays

l_hist(
  x,
  yshows = c("frequency", "density"),
  by = NULL,
  layout = c("grid", "wrap", "separate"),
  connectedScales = c("cross", "row", "column", "both", "x", "y", "none"),
  showStackedColors = TRUE,
  origin = NULL,
  binwidth = NULL,
  showBinHandle = FALSE,
  color = l_getOption("color"),
  active = TRUE,
  selected = FALSE,
  xlabel = NULL,
  parent = NULL,
  ...
)

Arguments

x
vector with numerical data to perform the binning on x,

yshows
one of "frequency" (default) or "density"

by
loon plot can be separated by some variables into multiple panels. This argument can take a vector, a list of same lengths or a data.frame as input.

layout
layout facets as 'grid', 'wrap' or 'separate'

connectedScales
Determines how the scales of the facets are to be connected depending on which layout is used. For each value of layout, the scales are connected as follows:

layout = "wrap": Across all facets, when connectedScales is

"x", then only the "x" scales are connected

"y", then only the "y" scales are connected

"both", both "x" and "y" scales are connected

"none", neither "x" nor "y" scales are connected. For any other value, only the "y" scale is connected.

layout = "grid": Across all facets, when connectedScales is

"cross", then only the scales in the same row and the same column are connected

"row", then both "x" and "y" scales of facets in the same row are connected

"column", then both "x" and "y" scales of facets in the same column are connected

"x", then all of the "x" scales are connected (regardless of column)

"y", then all of the "y" scales are connected (regardless of row)

"both", both "x" and "y" scales are connected in all facets

"none", neither "x" nor "y" scales are connected in any facets.

showStackedColors
if TRUE (default) then bars will be coloured according to colours of the points; if FALSE, then the bars will be a uniform colour except for highlighted points.

origin
numeric scalar to define the binning origin

binwidth
a numeric scalar to specify the binwidth If NULL binwidth is set using David Scott's rule when x is numeric (namely 3.49 * sd(x)/(n ^(1/3)) if sd(x) > 0 and 1 if sd(x) == 0) and using the minumum numerical difference between factor levels when x is a factor or a character vector (coerced to factor).

showBinHandle
If TRUE, then an interactive "bin handle" appears on the plot whose movement resets the origin and the binwidth. Default is FALSE

color
colour fills of bins; colours are repeated until matching the number x. Default is found using l_getOption("color").

active
a logical determining whether x appears or not (default is TRUE for all x). If a logical vector is given of length equal to the number of x, then it identifies which x appears (TRUE) and which does not (FALSE).

selected
a logical determining whether x appears selected at first (default is FALSE for all x). If a logical vector is given of length equal to the number of x, then it identifies which x is (TRUE) and which is not (FALSE).

xlabel
label to be used on the horizontal axis. If NULL, an attempt at a meaningful label inferred from x will be made.

parent
parent widget path

...
named arguments to modify the histogram plot states or layouts, see details.

Value

widget handle

Details

Note that when changing the yshows state from 'frequency' to 'density' you might have to use l_scaleto_world to show the complete histogram in the plotting region.

Some arguments to modify layouts can be passed through, e.g. "separate", "byrow", etc. Check l_facet to see how these arguments work.

For more information run: l_help("learn_R_display_hist")

See also

Examples

if(interactive()){

h <- l_hist(iris$Sepal.Length)

names(h)
h["xlabel"] <- "Sepal length"
h["showOutlines"] <- FALSE

h["yshows"]
h["yshows"] <- "density"
l_scaleto_plot(h)

h["showStackedColors"] <- TRUE
h['color'] <- iris$Species
h["showStackedColors"] <- FALSE
h["showOutlines"] <- TRUE
h["showGuides"] <- TRUE

# link another plot with the previous plot
h['linkingGroup'] <- "iris_data"
h2 <- with(iris, l_hist(Petal.Width,
                        linkingGroup="iris_data",
                        showStackedColors = TRUE))

# Get an R (grid) graphics plot of the current loon plot
plot(h)
# or with more control about grid parameters
grid.loon(h)
# or to save the grid data structure (grob) for later use
hg <- loonGrob(h)

}