Layer a smooth line for loon

Display a smooth line layer

l_layer_smooth(
  widget,
  x,
  y = NULL,
  method = "loess",
  group_by = "color",
  formula = y ~ x,
  se = TRUE,
  n = 80,
  span = 0.75,
  level = 0.95,
  method.args = list(),
  linecolor = "steelblue",
  se_color = "gray80",
  linewidth = 2,
  se_width = 4,
  label = "smooth",
  parent = "root",
  index = 0,
  ...
)

Arguments

widget
widget path name as a string

x
The x coordinates of line. If it is not provided, x will be inherited from widget

y
The y coordinates of line. If it is not provided, y will be inherited from widget

method
Smoothing method (function) to use, accepts either a character vector, e.g. "lm", "glm", "loess" or a function, e.g. MASS::rlm or mgcv::gam, stats::lm, or stats::loess.

group_by
Coordinates can be grouped by aesthetics attributes, e.g. "color". If x and y are inherited from widget, group_by can be set as any N dimensional states of corresponding widget; if not, group_by can only be set as either "linecolor" or "linewidth"

formula
Formula to use in smoothing function, eg. y ~ x, y ~ poly(x, 2), y ~ log(x)

se
Display confidence interval around smooth? (TRUE by default, see level to control.)

n
Number of points at which to evaluate smoother.

span
Controls the amount of smoothing for the default loess smoother. Smaller numbers produce wigglier lines, larger numbers produce smoother lines.

level
Level of confidence interval to use (0.95 by default).

method.args
List of additional arguments passed on to the modelling function defined by method.

linecolor
fitted line color. Note that the linecolor of fitted lines are respect to whether x and y are provided. If coordinates are inherited from widget, the linecolor can only be set via l_configure

se_color
fitted standard deviation polyline color

linewidth
fitted line width

se_width
fitted standard deviation polyline width

label
label used in the layers inspector

parent
group layer

index
index of the newly added layer in its parent group

...
additional state initialization arguments, see l_info_states

Examples

if(interactive()) {
p <- l_plot(iris, color = iris$Species)
# the fits are grouped by points color
l1 <- l_layer_smooth(p)
l_layer_hide(l1)
# the fitted line is based on all points
l2 <- l_layer_smooth(p, group_by = "")
l_layer_hide(l2)

# Draw a fitted line based on a new data set
shortSepalLength <- (iris$Sepal.Length < 5)
l3 <- l_layer_smooth(p,
                     x = iris$Sepal.Length[shortSepalLength],
                     y = iris$Sepal.Width[shortSepalLength],
                     method = "lm",
                     linecolor = "firebrick")

if(require(mgcv)) {
  p1 <- l_plot(iris)
  # a full tensor product smooth
  ## linecolor is the same with the points color
  l4 <- l_layer_smooth(p1,
                       method = "gam",
                       formula = y~te(x))
  l_layer_hide(l4)
  ## linecolor is set as the default input color 'steelblue'
  l5 <- l_layer_smooth(p1,
                       x = p1['x'],
                       y = p1['y'],
                       method = "gam",
                       formula = y~te(x))
}
}