2d Navigation Graph Setup with dynamic node fitering using a scatterplot matrix

Measures object is of class measures. When using measure objects then the measures can be dynamically re-calculated for a subset of the data.

# S3 method for measures
l_ng_plots(measures, ...)

Arguments

measures
object of class measures, see measures1d, measures2d.

...
arguments passed on to configure the scatterplot

Value

named list with plots-, graph-, plot-, navigator-, and context handle. The list also contains the environment of the the function call in env.

Details

Note that we provide the scagnostics2d function to create a measures object for the scagnostics measures.

For more information run: l_help("learn_R_display_graph.html#l_ng_plots")

See also

Examples

# NOT RUN {
# 2d measures
scags <- scagnostics2d(oliveAcids, separator='**')
scags()
ng <- l_ng_plots(scags, color=olive$Area)

# 1d measures
scale01 <- function(x){(x-min(x))/diff(range(x))}
m1d <- measures1d(sapply(iris[,-5], scale01),
     mean=mean, median=median, sd=sd,
     q1=function(x)as.vector(quantile(x, probs=0.25)),
     q3=function(x)as.vector(quantile(x, probs=0.75)))

m1d()

nav <- l_ng_plots(m1d, color=iris$Species)

# with only one measure
nav <- l_ng_plots(measures1d(oliveAcids, sd))

# with two measures
nav <- l_ng_plots(measures1d(oliveAcids, sd=sd, mean=mean))
# }