Draw a decomposed time series loon plot

l_plot_ts is a generic function for creating a decomposed time series plot. It is mainly used in l_plot.decomposed.ts and l_plot.stl

l_plot_ts(
  x,
  color = l_getOption("color"),
  size = l_getOption("size"),
  linecolor = l_getOption("color"),
  linewidth = l_getOption("linewidth"),
  xlabel = NULL,
  ylabel = NULL,
  title = NULL,
  tk_title = NULL,
  linkingGroup,
  showScales = TRUE,
  showGuides = TRUE,
  showLabels = TRUE,
  call = match.call(),
  ...
)

Arguments

x
Either an stl object or a decomposed.ts object.

color
points colour of all time series. Default is given by l_getOption("color").

size
points size of all time series. Default is given by l_getOption("size").

linecolor
line colour of all time series. Default is given by l_getOption("color").

linewidth
line width of all time series (incl. original and decomposed components. Default is given by l_getOption("linewidth").

xlabel
the labels for the x axes. This is a length four character vector one for each: of the original time series, the trend component, the seasonality component, and the remainder. If of length 1, the label is repeated; if NULL, xlabel is "time".

ylabel
the labels for the vertical axes. This is a length four character vector one for each: of the original time series, the trend component, the seasonality component, and the remainder. If NULL, the default, ylabel will be c("data", "trend", "seasonality", "remainder"); if a character vector of length 1, the label is repeated four times.

title
an overall title for the entire display. If NULL (the default), the title will be "Seasonal Trend Analysis".

tk_title
provides an alternative window name to Tk's wm title. If NULL, stl will be used.

linkingGroup
name of linking group. If missing, one is created from the data name and class associated with stlOrDecomposedTS.

showScales
a logical as to whether to display the scales on all axes, default is TRUE.

showGuides
a logical as to whether to display background guide lines on all plots, default is TRUE.

showLabels
a logical as to whether to display axes labels on all plots, default is TRUE.

call
a call in which all of the specified arguments are specified by their full names

...
keyword value pairs passed off to l_plot() which constructs each loon scatterplot component.

Value

A structure of class "l_ts" containing four loon plots each representing a part of the decomposition by name: "original", "trend", "seasonal", and "remainder".

See also