Model Prediction

It is entirely for the purpose of plotting fits and intervals on a scatterplot (or histogram). It is a generic function to predict models for loon smooth layer (a wrap of the function predict). However, the output is unified.

l_predict(model, ...)

# S3 method for default
l_predict(model, ...)

# S3 method for lm
l_predict(
  model,
  newdata = NULL,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)

# S3 method for nls
l_predict(
  model,
  newdata = NULL,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)

# S3 method for glm
l_predict(
  model,
  newdata = NULL,
  interval = c("none", "confidence"),
  level = 0.95,
  ...
)

# S3 method for loess
l_predict(
  model,
  newdata = NULL,
  interval = c("none", "confidence", "prediction"),
  level = 0.95,
  ...
)

Arguments

model
a model object for which prediction is desired

...
arguments passed in predict

newdata
optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.

interval
type of interval, could be "none", "confidence" or "prediction" (not for glm)

level
confidence level

Value

A data frame is returned with x (if newdata is given) and y. If the interval is not none, two more columns, lower (lower interval) and upper (upper interval) will be returned.

Examples

y <- rnorm(10)
x <- rnorm(10)
model1 <- lm(y ~ x)
# formal output
pre <- l_predict(model1, newdata = data.frame(x = sort(x)),
                 interval = "conf")
head(pre)

#>            x         y      lower     upper
#> 1 -1.3189259 0.5765009 -0.1009026 1.2539045
#> 2 -1.2895577 0.5628749 -0.1002272 1.2259771
#> 3 -0.9818526 0.4201082 -0.1080509 0.9482673
#> 4 -0.9211751 0.3919555 -0.1139075 0.8978185
#> 5 -0.5851407 0.2360447 -0.1879557 0.6600452
#> 6 -0.3166412 0.1114683 -0.3148288 0.5377654

if(interactive()) {
p <- with(cars, l_plot(speed, dist))
l_predict.smooth.spline <- function(model, interval = c("confidence", "none"),
                                    level = 0.95, ...) {
# confidence interval of `smooth.spline`
# https://stackoverflow.com/questions/23852505/how-to-get-confidence-interval-for-smooth-spline
  interval <- match.arg(interval)

  res <- (model$yin - model$y)/(1 - model$lev)     # jackknife residuals
  sigma <- sqrt(var(res))                          # estimate sd
  std <- stats::qnorm(level / 2 + 0.5)
  upper <- model$y + std * sigma * sqrt(model$lev) # upper 95% conf. band
  lower <- model$y - std * sigma * sqrt(model$lev) # lower 95% conf. band

  data.frame(y = model$yin, lower = lower, upper = upper)
}
l <- l_layer_smooth(p, method = "smooth.spline", interval = "confidence")
}