Modify States of a Plot that are Linked in Loon's Standard Linking Model

Loon's standard linking model is based on three levels, the linkingGroup and linkingKey states and the used linkable states. See the details below.

l_setLinkedStates(widget, states)

Arguments

widget
widget path as a string or as an object handle

states
used linkable state names, see in details below

Details

Loon's standard linking model is based on two states, linkingGroup and linkingKey. The full capabilities of the standard linking model are described here. However, setting the linkingGroup states for two or more displays to the same string is generally all that is needed for linking displays that plot data from the same data frame. Changing the linking group of a display is also the only linking-related action available on the analysis inspectors.

The first linking level is as follows: loon's displays are linked if they share the same string in their linkingGroup state. The default linking group 'none' is a keyword and leaves a display un-linked.

The second linking level is as follows. All n-dimensional states can be linked between displays. We call these states linkable. Further, only linkable states with the same name can be linked between displays. One consequence of this shared state name rule is that, with the standard linking model, the linewidth state of a serialaxes display cannot be linked with the size state of a scatterplot display. Also, each display maintains a list that defines which of its linkable states should be used for linking; we call these states the used linkable states. The default used linkable states are as follows

Display	Default

used linkable	states

scatterplot

selected	,

color	,

active	,

size

histogram

selected

,	color

,

active

serialaxes

selected	,

color	,

active

graph

selected	,

color	,

active	,

size

If any two displays are set to be linked (i.e. they share the same linking group) then the intersection of their used linkable states are actually linked.

The third linking level is as follows. Every display has a n-dimensional linkingKey state. Hence, every data point has an associated linking key. Data points between linked plots are linked if they share the same linking key.