Open a browser with loon's R documentation webpage

l_web opens a browser with the relevant page on the official loon documentation website. This is constructed by joining together the information provided by the arguments site/package/directory/page.

Default would be the documentation found at https://great-northern-diver.github.io/loon/.

l_web(
  package = c("loon", "loon.data", "loon.ggplot", "loon.tourr", "ggmulti", "zenplots",
    "diveR"),
  page = "index",
  directory = c("home", "reference", "articles"),
  site = "https://great-northern-diver.github.io",
  ...
)

Arguments

package
a string identifying the package name having an online documentation (default "loon").

page
relative path to a page (the ".html" part may be omitted)

directory
if "home" (the default) then page is ignored and the browser will open at the home page of the official documentation website. If page refers to a package manual reference, then directory must be "reference"; if page refers to the name of a vignette file, then directory should be "articles"

site
the URL of the site (default "https://great-northern-diver.github.io") prefixing the path to the requested documentation.

...
arguments forwarded to browseURL(), e.g. to specify a browser

See also

Examples

if (FALSE) {
l_web()
#
vignette("introduction", package = "loon")
# or
l_web(page = "introduction", directory = "articles")
# or
l_web(package = "loon.data", directory = "reference")
#
help(l_hist)
l_web(page = "l_hist", directory = "reference")
}